move ChangeDayButton to its own file
